Refactor galaxy.datatypes.checkers into galaxy.util.checkers.
 - Primary reason is because it is used all over the place in contexts that have nothing to do with Galaxy datatype classes or datatype registery or sniffing, most importantly by the code in the tool shed for determining if something is a tool or not.
 - This code has no dependencies on the datatypes module or any hard dependencies on things not in the Python standard library.
