+++ lib/galaxy/datatypes/sniff.py
@@ -15,7 +15,7 @@ from encodings import search_function as encodings_search_function
 
 from galaxy import util
 from galaxy.util import multi_byte
-from galaxy.datatypes.checkers import check_binary, check_html, is_gzip
+from galaxy.util.checkers import check_binary, check_html, is_gzip
 from galaxy.datatypes.binary import Binary
 
 log = logging.getLogger(__name__)
