+++ lib/galaxy/datatypes/tabular.py
@@ -13,7 +13,7 @@ from cgi import escape
 
 from galaxy import util
 from galaxy.datatypes import data, metadata
-from galaxy.datatypes.checkers import is_gzip
+from galaxy.util.checkers import is_gzip
 from galaxy.datatypes.metadata import MetadataElement
 from galaxy.datatypes.sniff import get_headers
 from galaxy.util.json import dumps
