+++ lib/galaxy/util/checkers.py
