+++ lib/galaxy/webapps/tool_shed/controllers/upload.py
@@ -7,7 +7,7 @@ import urllib
 
 from galaxy import util
 from galaxy import web
-from galaxy.datatypes import checkers
+from galaxy.util import checkers
 from galaxy.web.base.controller import BaseUIController
 from tool_shed.util.web_util import escape
