+++ lib/tool_shed/util/commit_util.py
@@ -9,7 +9,7 @@ import tempfile
 from sqlalchemy.sql.expression import null
 
 import tool_shed.repository_types.util as rt_util
-from galaxy.datatypes import checkers
+from galaxy.util import checkers
 from tool_shed.tools import data_table_manager
 from tool_shed.util import basic_util, hg_util, shed_util_common as suc
