+++ tools/data_source/upload.py
@@ -18,11 +18,11 @@ from galaxy import util
 import galaxy.model  # noqa
 from galaxy.datatypes import sniff
 from galaxy.datatypes.binary import Binary
-from galaxy.datatypes.checkers import check_binary, check_bz2, check_gzip, check_html, check_image, check_zip
 from galaxy.datatypes.registry import Registry
 from galaxy.datatypes.util.image_util import get_image_ext
 from galaxy.util.json import dumps, loads
 from galaxy.util import multi_byte
+from galaxy.util.checkers import check_binary, check_bz2, check_gzip, check_html, check_image, check_zip
 
 try:
     import Image as PIL
